I corrected my Task1 (I divided the yml into three different tasks to make it easier to check. I also divided my project into two files and divided the tests as well.)
