--- .github/workflows/code_style.yml
@@ -1,64 +1,0 @@
-name: Check code style and doing pytest, mypy
-
-on: [push, pull_request]
-
-jobs:
-  style:
-    runs-on: ubuntu-latest
-    strategy:
-      matrix:
-        python-version: ['3.8']
-    steps:
-      - name: Set up Git repository
-        uses: actions/checkout@v2
-      - name: Set up Python ${{ matrix.python-version }}
-        uses: actions/setup-python@v2
-        with:
-          python-version: ${{ matrix.python-version }}
-      - name: Install requirements
-        run: |
-          python -m pip install --upgrade pip wheel setuptools
-          python -m pip install -r requirements.txt
-          python -m pip list
-      - name: Install pre-commit
-        run: |
-          pre-commit install
-      - name: Run pre-commit
-        run: |
-          pre-commit run --all-files --color always --verbose --show-diff-on-failure
-
-  test:
-    runs-on: ubuntu-latest
-    steps:
-      - name: Checkout code
-        uses: actions/checkout@v2
-      - name: Set up Python
-        uses: actions/setup-python@v2
-        with:
-          python-version: '3.8'
-
-      - name: Install dependencies
-        run: |
-          python -m pip install --upgrade pip
-          pip install -r requirements.txt
-          pip install pytest
-
-      - name: Run tests
-        run: |
-          python scripts/run_tests.py
-
-  type-check:
-    runs-on: ubuntu-latest
-    steps:
-      - name: Checkout code
-        uses: actions/checkout@v2
-      - name: Set up Python
-        uses: actions/setup-python@v2
-        with:
-          python-version: '3.8'
-      - name: Install mypy
-        run: |
-          pip install mypy
-      - name: Check types
-        run: |
-          mypy project/
