+++ .github/workflows/code_style_check.yml
@@ -1,0 +1,33 @@
+name: Code Style Check
+
+on: [push, pull_request]
+
+jobs:
+  style:
+    runs-on: ubuntu-latest
+    strategy:
+      matrix:
+        python-version: ['3.8']
+    
+    steps:
+      - name: Set up Git repository
+        uses: actions/checkout@v2
+      
+      - name: Set up Python ${{ matrix.python-version }}
+        uses: actions/setup-python@v2
+        with:
+          python-version: ${{ matrix.python-version }}
+      
+      - name: Install requirements
+        run: |
+          python -m pip install --upgrade pip wheel setuptools
+          python -m pip install -r requirements.txt
+          python -m pip list
+      
+      - name: Install pre-commit
+        run: |
+          pre-commit install
+      
+      - name: Run pre-commit
+        run: |
+          pre-commit run --all-files
