+++ .github/workflows/type_check.yml
@@ -1,0 +1,24 @@
+name: Type Check
+
+on: [push, pull_request]
+
+jobs:
+  type-check:
+    runs-on: ubuntu-latest
+    
+    steps:
+      - name: Checkout code
+        uses: actions/checkout@v2
+      
+      - name: Set up Python
+        uses: actions/setup-python@v2
+        with:
+          python-version: '3.8'
+      
+      - name: Install mypy
+        run: |
+          pip install mypy
+      
+      - name: Check types
+        run: |
+          mypy project/
