+++ project/matrixs.py
@@ -1,0 +1,32 @@
+from typing import List, Union, Any
+import math
+
+# This function add matrix1 and matrix2
+def matrix_addition(matrix1: List[List[Union[int, float]]], matrix2: List[List[Union[int, float]]]) -> List[List[Union[int, float]]]:
+
+    if len(matrix1) != len(matrix2) or any( len(row1) != len(row2) for row1, row2 in zip(matrix1, matrix2)):
+        raise ValueError("The matrices must be the same size")
+
+    return [[x + y for x, y in zip(row1, row2)] for row1, row2 in zip(matrix1, matrix2)]
+
+
+# This function multiplies matrix1 and matrix2
+def matrix_multiplication(matrix1: List[List[Union[int, float]]], matrix2: List[List[Union[int, float]]]) -> List[List[Union[int, float]]]:
+
+    if len(matrix1[0]) != len(matrix2):
+        raise ValueError("the number of columns in a matrix must be equal to the number of rows in another matrix")
+
+    result = []
+    for i in range(len(matrix1)):
+        row = []
+        for j in range(len(matrix2[0])):
+            element = sum(matrix1[i][k] * matrix2[k][j] for k in range(len(matrix2)))
+            row.append(element)
+        result.append(row)
+
+    return result
+
+
+# This function transpose matrix
+def matrix_transpose(matrix: List[List[Any]]) -> List[List[Any]]:
+    return [list(row) for row in zip(*matrix)]  # '*' - unpacks the matrix
